Index interpretations
Has index interpretation logic implemented to enable the user to easily
understand what the numbers mean

+++ semantiQ/app/readibility/IndexARI.java
@@ -6,7 +6,7 @@ public class IndexARI implements IReadabilityIndex {
 	
 	public void calculate(int totalCharactersInText, int totalWords, int totalSentences) {
 		
-		rating = (double) (4.71 * (totalCharactersInText/totalWords) + 0.5 * (totalWords/totalSentences) - 21.43);
+		rating = (4.71 * (totalCharactersInText/(double)totalWords) + 0.5 * (totalWords/(double)totalSentences) - 21.43);
 	}
 	
 	@Override
